Remove env-inject plugin dependency.
